Ask for a code at GA setup or GA disable

+++ application/controllers/user.php
@@ -2,6 +2,7 @@
 namespace application\controllers;
 use \library as h;
 use \library\MVC as l;
+use \Muonium\GoogleAuthenticator as ga;
 use \application\models as m;
 
 class user extends l\Controller {
@@ -160,17 +161,34 @@ class user extends l\Controller {
 			$resp['token'] = $this->_token;
 	        $this->_modelUser = new m\Users($this->_uid);
 	        $s = 0;
+            $isValid = true;
+            
 	        if(isset($data->doubleAuth) && ($data->doubleAuth === 1 || $data->doubleAuth === 2)) {
 				$s = $data->doubleAuth;
-                if($this->_modelUser->isDoubleAuthGA()) {
-                    $this->_modelUser->deleteBackupCodes();
-                    $this->_modelUser->deleteSecretKey();
+                if($this->_modelUser->isDoubleAuthGA() || $data->doubleAuth === 2) {
+                    $googleAuth = new ga\GoogleAuthenticator();
+                    $secret = $this->_modelUser->getSecretKeyGA();
+                    
+                    if(!($googleAuth->checkCode($secret, $data->code))) {
+                        $isValid = false;
+                    }
+                    
+                    if($this->_modelUser->isDoubleAuthGA()) {
+                        $this->_modelUser->deleteBackupCodes();
+                        $this->_modelUser->deleteSecretKey();
+                    }
                 }
 			}
-	        if($this->_modelUser->updateDoubleAuth($s)) {
-				$resp['code'] = 200;
-				$resp['status'] = 'success';
-	        }
+            
+            if(!($isValid)) {
+                $resp['code'] = 403;
+				$resp['status'] = 'badCode';
+            } else {
+                if($this->_modelUser->updateDoubleAuth($s)) {
+				    $resp['code'] = 200;
+				    $resp['status'] = 'success';
+	           }
+            }
 		}
 
 		http_response_code($resp['code']);
